requirements.txt
openai>=1.0.0
psycopg2-binary
tiktoken
streamlit
tqdm
python-dotenv
pgvector
sqlalchemy
pandas
numpy
rich
plotly 
scikit-learn